+++ matgpt/src/main/java/com/ktc/matgpt/review/dto/ReviewResponse.java
@@ -48,7 +48,7 @@ public class ReviewResponse {
         private int averageCostPerPerson;
         private int peopleCount;
         private String createdAt;
-        private double rating;
+        private int rating;
         private int recommendCount;
         private String content;
         private List<FindByReviewIdDTO.ImageDTO> reviewImages;
@@ -103,7 +103,7 @@ public class ReviewResponse {
                 private String name;
                 private double location_x;
                 private double location_y;
-                private double rating;
+                private int rating;
 
                 public TagDTO(Tag tag) {
                     this.name = tag.getFood().getFoodName();
@@ -149,7 +149,7 @@ public class ReviewResponse {
         @ToString
         public static class StoreReviewDTO {
             private Long reviewId;
-            private double rating;
+            private int rating;
             private String content;
             private LocalDateTime createdAt;
             private List<String> imageUrls;
@@ -208,7 +208,7 @@ public class ReviewResponse {
         @ToString
         public static class UserReviewDTO {
             private Long id;
-            private double rating;
+            private int rating;
             private String content;
             private LocalDateTime createdAt;
             private String storeImage;
